Refactored custom-code feature into `customcode` package
This is to prepare for the new readers/writers we'll be adding in 2.2. No change in functionality, just moving classes and tests.
