+++ src/main/java/com/marklogic/spark/reader/customcode/CustomCodePartitionReaderFactory.java
@@ -1,6 +1,5 @@
-package com.marklogic.spark.reader;
+package com.marklogic.spark.reader.customcode;
 
-import com.marklogic.spark.CustomCodeContext;
 import org.apache.spark.sql.catalyst.InternalRow;
 import org.apache.spark.sql.connector.read.InputPartition;
 import org.apache.spark.sql.connector.read.PartitionReader;
