+++ src/test/java/com/marklogic/spark/writer/customcode/ProcessMultipleRowsWithCustomCodeTest.java
@@ -1,7 +1,8 @@
-package com.marklogic.spark.writer;
+package com.marklogic.spark.writer.customcode;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.marklogic.spark.Options;
+import com.marklogic.spark.writer.AbstractWriteTest;
 import org.apache.spark.sql.SaveMode;
 import org.junit.jupiter.api.Test;
